requirements.txt
pandas>=1.5.0
numpy>=1.21.0
requests>=2.28.0
openai>=1.0.0
google-generativeai>=0.3.0
streamlit>=1.28.0
plotly>=5.15.0
scipy>=1.10.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
faker>=18.0.0